+++ src/utils/postprocessing-splitters/index.ts
@@ -1,0 +1,1 @@
+// TODO: !!!! Export here all splitter postprocessing functions and use them in same way as custom postrocessing functions