+++ src/utils/postprocessing-splitters/test/assertsSplitterFunctionIsWorking.ts
@@ -1,0 +1,9 @@
+export function assertsSplitterFunctionIsWorking(){
+  // TODO: !!!! Implement
+}
+
+/**
+ * TODO: !!!! Use
+ * TODO: !!!! Export to utils
+ * TODO: !!!! Annotate
+ */