+++ src/utils/postprocessing/test/postprocessing.test.ts
@@ -1,0 +1,7 @@
+import { describe, it } from '@jest/globals';
+
+describe('how postprocessing functions works', () => {
+    it('should work with foo', () => {
+        expect('!!!!').toBe('!!!!');
+    });
+});